Rename internal files
