+++ src/Label.js
