+++ src/index.js
@@ -1,2 +1,2 @@
-import ReactMinimalPieChart from './ReactMinimalPieChart';
-export default ReactMinimalPieChart;
+import Chart from './Chart';
+export default Chart;
